updated something

+++ data/SparseTable.cpp
@@ -19,7 +19,7 @@ public:
     vector<vector<T>> mat;
     SparseTableChan(const vector<T> &a)
     {
-        n = static_cast<int>(a.size());
+        n = a.size();
         int max_log = 32 - __builtin_clz(n);
         mat.resize(max_log);
         mat[0] = a;
